Bullet Collision, damage, azimuth fix
Added bullet Collision and particle effects. Bullets deal damage when
hitting players. Ball Spawn was moved infront of player as I was unable
to tag bullets with the player who spawned them to avoid self collision.
Azimuth was improved, it now uses x and y. I have code for it to follow
cursor but weird bug when loading world.